+++ src/utils/logger.ts
@@ -1,6 +1,6 @@
-import GitTextUtils from './gitTextUtils';
 import { processLog } from '../extension';
 import { MagitProcessLogEntry } from '../models/magitProcessLogEntry';
+import { IExecutionResult } from '../common/gitApiExtensions';
 
 export default class MagitLogger {
 
@@ -8,6 +8,10 @@ export default class MagitLogger {
     const logEntry = { command: ['git', ...args], index: processLog.length };
     processLog.push(logEntry);
     return logEntry;
+  }
+
+  public static logGitResult(result: IExecutionResult<string>, entry: MagitProcessLogEntry) {
+    entry.output = result.stdout;
   }
 
   public static logGitError(error: any, entry: MagitProcessLogEntry) {