+++ src/views/dispatchView.ts
@@ -19,7 +19,7 @@ export class DispatchView extends DocumentView {
     m Merging           M Remoting         t Tagging
     P Pushing           r Rebasing         X Resetting
     y Show Refs         z Stashing         % Worktree
-    i Ignoring`));
+    i Ignoring                             ! Running`));
   }
 
   public update(state: MagitState): void { }